pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memories-dev"
version = "1.0.5"
description = "A package for daily synthesis of Earth Memories, helping agents answer about location-first queries."
readme = "README.md"
authors = [{ name = "Memories-dev", email = "hello@memories-dev.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Natural Language :: English"
]
keywords = ["earth", "memory", "ai", "geo", "satellite"]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=1.3.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "duckdb>=0.9.0",
    "matplotlib>=3.7.0",
    "ipywidgets>=8.0.0",
    "pyarrow>=14.0.1",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.3.0",
    "rasterio>=1.3.8",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "opencv-python>=4.8.0",
    "albumentations>=1.3.1",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "geopy>=2.3.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "mercantile>=1.2.1",
    "mapbox-vector-tile>=2.0.1",
    "pyproj>=3.6.1",
    "pystac>=1.8.0",
    "xarray>=2023.0.0",
    "dask>=2024.1.0",
    "dask[array]>=2024.1.0",
    "spacy>=3.8.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-benchmark>=4.0.0",
    "pytest-timeout>=2.2.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "nbsphinx>=0.9.0",
    "pandoc>=2.3"
]
requires-python = ">=3.9"

[project.optional-dependencies]
gpu = [
    "faiss-gpu>=1.7.4",
    "cupy-cuda12x>=12.0.0",
    "cudf>=24.2.0",
    "cuspatial>=24.2.0"
]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-benchmark>=4.0.0",
    "pytest-timeout>=2.2.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0"
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "nbsphinx>=0.9.0",
    "pandoc>=2.3"
]

[project.urls]
Homepage = "https://github.com/Vortx-AI/memories-dev"
Documentation = "https://docs.memories-dev.dev"
Repository = "https://github.com/Vortx-AI/memories-dev.git"
Issues = "https://github.com/Vortx-AI/memories-dev/issues"
Changelog = "https://github.com/Vortx-AI/memories-dev/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["memories"]
package-dir = {"memories" = "memories"}

[tool.setuptools.package-data]
memories = [
    "config/*.yaml",
    "data/*.json",
    "models/config/*.json",
    "utils/styles/*.json"
]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra -q --cov=memories"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = ["asyncio: mark test as async"] 